moved number utility method to a new utility class
