--- application/src/test/java/org/thingsboard/server/utils/NumberUtilsTest.java
@@ -1,32 +1,0 @@
-/**
- * Copyright © 2016-2026 The Thingsboard Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.thingsboard.server.utils;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class NumberUtilsTest {
-
-    private final double doubleVal = 1729.1729;
-
-    @Test
-    public void roundResult() {
-        Assertions.assertEquals(1729.1729, NumberUtils.roundResult(doubleVal, null));
-        Assertions.assertEquals(1729, NumberUtils.roundResult(doubleVal, 0));
-        Assertions.assertEquals(1729.17, NumberUtils.roundResult(doubleVal, 2));
-    }
-
-}
