+++ common/script/script-api/src/test/java/org/thingsboard/script/api/tbel/TbUtilsTest.java
@@ -280,13 +280,6 @@ public class TbUtilsTest {
     }
 
     @Test
-    public void toFixedFloat() {
-        float actualF = TbUtils.toFixed(floatVal, 3);
-        Assertions.assertEquals(1, Float.compare(floatVal, actualF));
-        Assertions.assertEquals(0, Float.compare(29.298f, actualF));
-    }
-
-    @Test
     public void parseBytesToFloat() {
         byte[] floatValByte = {0x0A};
         Assertions.assertEquals(0, Float.compare(1.4E-44f, TbUtils.parseBytesToFloat(floatValByte)));
@@ -459,13 +452,6 @@ public class TbUtilsTest {
         Assertions.assertEquals(0, Double.compare(doubleValRev, TbUtils.parseHexToDouble(longValHex, false)));
         Assertions.assertEquals(0, Double.compare(doubleVal, TbUtils.parseBigEndianHexToDouble(longValHex)));
         Assertions.assertEquals(0, Double.compare(doubleVal, TbUtils.parseLittleEndianHexToDouble(longValHexRev)));
-    }
-
-    @Test
-    public void toFixedDouble() {
-        double actualD = TbUtils.toFixed(doubleVal, 3);
-        Assertions.assertEquals(-1, Double.compare(doubleVal, actualD));
-        Assertions.assertEquals(0, Double.compare(1729.173, actualD));
     }
 
     @Test
@@ -1145,19 +1131,6 @@ public class TbUtilsTest {
         expected = "BB53";
         actual = TbUtils.bytesToHex(expectedList);
         Assertions.assertEquals(expected, actual);
-    }
-
-    @Test
-    public void toInt() {
-        Assertions.assertEquals(1729, TbUtils.toInt(doubleVal));
-        Assertions.assertEquals(13, TbUtils.toInt(12.8));
-        Assertions.assertEquals(28, TbUtils.toInt(28.0));
-    }
-
-    @Test
-    public void isNaN() {
-        assertFalse(TbUtils.isNaN(doubleVal));
-        assertTrue(TbUtils.isNaN(Double.NaN));
     }
 
     @Test
